backend/pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "rag-chatbot-backend"
version = "0.1.0"
description = "Backend service for the Integrated RAG Chatbot."
requires-python = ">=3.11"
dependencies = [
    "fastapi>=0.109.0",
    "uvicorn[standard]>=0.25.0",
    "pydantic>=2.5.3",
    "python-dotenv>=1.0.0",
    "qdrant-client>=1.7.0",
    "asyncpg>=0.29.0",
    "google-generativeai>=0.5.4",
    "openai>=1.55.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0.0",
]

[tool.hatch.build.targets.wheel]
packages = ["src"]
